Added Feed forward Networks and Decision Trees
Genetic Algorithms are too confusing, haha. I will work on trying to get that to happen instead, but it might be difficult. I was considering having each generation be 2 big, and having each pop. member play against each other. Have to do more research on how to code the GA, wrapping my head around how it 100% works will be key. 
For now, the game works, and the game with networks built in tries very hard to do stuff, but is impractical in terms of performance. I think this is also due to a lack of tree depth.